backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
fastf1==3.3.5
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
structlog==24.1.0
slowapi==0.1.9
google-cloud-storage==2.14.0
python-json-logger==2.0.7
tenacity==8.2.3
gunicorn==21.2.0
alembic>=1.12.1
python-dotenv>=1.0.0
fastapi-utils>=0.2.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
matplotlib>=3.7.0 